requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
python-dateutil==2.9.0.post0
tweepy==4.14.0
